+++ shared/ui/src/commonMain/kotlin/com/shub39/grit/shared/ui/habit/ui/sections/Calendar.kt
@@ -212,6 +212,7 @@ private fun YearlyCalendar(
                     doneDates = doneDates,
                     today = today,
                     habitDays = currentHabit.habit.days,
+                    startDate = currentHabit.habit.time.date,
                     edgeWeeks = edgeWeeks,
                     onDateClick = { onDateClick(currentHabit.habit, it) },
                 )
@@ -250,6 +251,7 @@ private fun MonthlyCalendar(
                 doneDates = doneDates,
                 today = today,
                 habitDays = currentHabit.habit.days,
+                startDate = currentHabit.habit.time.date,
                 edgeWeeks = edgeWeeks,
                 onDateClick = { onDateClick(currentHabit.habit, it) },
             )